# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: BookWise API
  description: >-
    Contrato da API do BookWise - sistema de gestao de bibliotecas e livrarias
    fisicas e digitais. Este arquivo e o **contrato de referencia (design)** do
    projeto. Os campos e recursos aqui sao provisorios e serao ajustados quando
    o MER/DER final for definido.

    Abordagem: code-first. O contrato vivo, gerado a partir das anotacoes do
    codigo (springdoc), fica em `/v3/api-docs`. Este arquivo serve como guia
    para o time se basear ao implementar.
  version: 0.0.1
  contact:
    name: Equipe BookWise
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Ambiente local
...
paths:
...
  /api/v1/reservations:
    get:
      tags: [Reservations]
      summary: Lista reservas
      operationId: listReservations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: Pagina de reservas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationPage'
    post:
      tags: [Reservations]
      summary: Registra uma reserva
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequest'
      responses:
        '201':
          description: Reserva registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/reservations/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Reservations]
      summary: Busca reserva por id
      operationId: getReservationById
      responses:
        '200':
          description: Reserva encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/reservations/{id}/cancel:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      tags: [Reservations]
      summary: Cancela uma reserva
      operationId: cancelReservation
      responses:
        '200':
          description: Reserva cancelada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================================================================
  # FINES
  # ==================================================================
  /api/v1/fines:
    get:
      tags: [Fines]
      summary: Lista multas
      operationId: listFines
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: Pagina de multas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinePage'
  /api/v1/fines/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Fines]
      summary: Busca multa por id
      operationId: getFineById
      responses:
        '200':
          description: Multa encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fine'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/fines/{id}/pay:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      tags: [Fines]
      summary: Registra o pagamento de uma multa
      operationId: payFine
      responses:
        '200':
          description: Multa paga
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fine'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================================================================
  # USERS
  # ==================================================================
...
  /api/v1/books:
    get:
      tags: [Books]
      summary: Lista livros
      operationId: listBooks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: q
          in: query
          description: Filtro por titulo, autor ou ISBN
          schema:
            type: string
      responses:
        '200':
          description: Pagina de livros
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookPage'
    post:
      tags: [Books]
      summary: Cria um livro
      operationId: createBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
      responses:
        '201':
          description: Livro criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/books/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Books]
      summary: Busca livro por id
      operationId: getBookById
      responses:
        '200':
          description: Livro encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Books]
      summary: Atualiza um livro
      operationId: updateBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
      responses:
        '200':
          description: Livro atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Books]
      summary: Remove um livro
      operationId: deleteBook
      responses:
        '204':
          description: Livro removido
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================================================================
  # LOANS
  # ==================================================================
  /api/v1/loans:
    get:
      tags: [Loans]
      summary: Lista emprestimos
      operationId: listLoans
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: Pagina de emprestimos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanPage'
    post:
      tags: [Loans]
      summary: Registra um emprestimo
      operationId: createLoan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanRequest'
      responses:
        '201':
          description: Emprestimo registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/loans/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Loans]
      summary: Busca emprestimo por id
      operationId: getLoanById
      responses:
        '200':
          description: Emprestimo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/loans/{id}/return:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      tags: [Loans]
      summary: Registra a devolucao de um emprestimo
      operationId: returnLoan
      responses:
        '200':
          description: Devolucao registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================================================================
  # SALES
  # ==================================================================
  /api/v1/sales:
    get:
      tags: [Sales]
      summary: Lista vendas
      operationId: listSales
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: Pagina de vendas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalePage'
    post:
      tags: [Sales]
      summary: Registra uma venda
      operationId: createSale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleRequest'
      responses:
        '201':
          description: Venda registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/sales/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Sales]
      summary: Busca venda por id
      operationId: getSaleById
      responses:
        '200':
          description: Venda encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/sales/{id}/cancel:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      tags: [Sales]
      summary: Cancela uma venda
      operationId: cancelSale
      responses:
        '200':
          description: Venda cancelada (estoque restaurado)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Identificador do recurso
      schema:
        type: integer
        format: int64
    PageParam:
      name: page
      in: query
      description: Numero da pagina (base 0)
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      description: Quantidade de itens por pagina
      schema:
        type: integer
        format: int32
        default: 20
        minimum: 1
        maximum: 100
    UserIdParam:
      name: userId
      in: query
      required: false
      description: Filtra os registros pelo identificador do usuario
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: Requisicao invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso nao encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    BookFormat:
      type: string
      description: Formato do livro
      enum: [PHYSICAL, DIGITAL]
    LoanStatus:
      type: string
      description: Situacao do emprestimo
      enum: [ACTIVE, RETURNED, LATE]
    SaleStatus:
      type: string
      description: Situacao da venda
      enum: [PENDING, PAID, CANCELLED]
    ReservationStatus:
      type: string
      description: Situacao da reserva
      enum: [ACTIVE, FULFILLED, CANCELLED, EXPIRED]
    FinePaymentStatus:
      type: string
      description: Situacao de pagamento da multa
      enum: [PENDING, PAID]

    # ---------- User ----------
...
    BookRequest:
      type: object
      required: [title, author, isbn, format]
      properties:
        title:
          type: string
          example: Clean Code
        author:
          type: string
          example: Robert C. Martin
        isbn:
          type: string
          example: "9780132350884"
        genre:
          type: string
          maxLength: 50
          example: Tecnologia
        publishedYear:
          type: integer
          format: int32
          example: 2008
        format:
          $ref: '#/components/schemas/BookFormat'
        price:
          type: number
          format: double
          example: 89.90
        stock:
          type: integer
          format: int32
          description: Quantidade em estoque (para livros fisicos)
          example: 10
        categoryIds:
          type: array
          description: Ids das categorias associadas (N:N)
          items:
            type: integer
            format: int64
          example: [1, 4]
    Book:
      allOf:
        - $ref: '#/components/schemas/BookRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
              example: 1
            categories:
              type: array
              description: Categorias associadas ao livro
              items:
                $ref: '#/components/schemas/CategoryRef'
            createdAt:
              type: string
              format: date-time
    CategoryRef:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: Tecnologia

    # ---------- Loan ----------
    LoanItemRequest:
      type: object
      required: [bookId]
      properties:
        bookId:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          format: int32
          minimum: 1
          default: 1
          example: 1
    LoanRequest:
      type: object
      required: [userId, items]
      properties:
        userId:
          type: integer
          format: int64
          example: 1
        dueDate:
          type: string
          format: date
          description: "Data prevista de devolucao (padrao: hoje + 14 dias)"
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LoanItemRequest'
    LoanItem:
      type: object
      properties:
        bookId:
          type: integer
          format: int64
        bookTitle:
          type: string
          example: Clean Code
        quantity:
          type: integer
          format: int32
          example: 1
    Loan:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: integer
          format: int64
        userName:
          type: string
          example: Maria Silva
        items:
          type: array
          items:
            $ref: '#/components/schemas/LoanItem'
        loanDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        returnDate:
          type: string
          format: date
          nullable: true
        status:
          $ref: '#/components/schemas/LoanStatus'

    # ---------- Sale ----------
    SaleItemRequest:
      type: object
      required: [bookId]
      properties:
        bookId:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          format: int32
          minimum: 1
          default: 1
          example: 1
        unitPrice:
          type: number
          format: double
          description: "Preco unitario praticado (padrao: preco do livro)"
          example: 89.90
    SaleRequest:
      type: object
      required: [userId, items]
      properties:
        userId:
          type: integer
          format: int64
          example: 1
        paymentMethod:
          type: string
          example: PIX
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SaleItemRequest'
    SaleItem:
      type: object
      properties:
        bookId:
          type: integer
          format: int64
        bookTitle:
          type: string
          example: Clean Code
        quantity:
          type: integer
          format: int32
          example: 1
        unitPrice:
          type: number
          format: double
          example: 89.90
    Sale:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: integer
          format: int64
        userName:
          type: string
          example: Maria Silva
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleItem'
        paymentMethod:
          type: string
          example: PIX
        saleDate:
          type: string
          format: date
        totalPrice:
          type: number
          format: double
          example: 89.90
        status:
          $ref: '#/components/schemas/SaleStatus'

    # ---------- Category ----------
...
    PageMeta:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    BookPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Book'
        meta:
          $ref: '#/components/schemas/PageMeta'
    LoanPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Loan'
        meta:
          $ref: '#/components/schemas/PageMeta'
    SalePage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Sale'
        meta:
          $ref: '#/components/schemas/PageMeta'
...
    ReservationRequest:
      type: object
      required: [userId, bookId]
      properties:
        userId:
          type: integer
          format: int64
          example: 1
        bookId:
          type: integer
          format: int64
          example: 1
        expirationDate:
          type: string
          format: date
          description: "Data de expiracao (padrao: hoje + 7 dias)"
    Reservation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        userName:
          type: string
          example: Joao Souza
        bookId:
          type: integer
          format: int64
        bookTitle:
          type: string
          example: O Senhor dos Aneis
        reserveDate:
          type: string
          format: date
        expirationDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/ReservationStatus'
    ReservationPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Reservation'
        meta:
          $ref: '#/components/schemas/PageMeta'

    # ---------- Fine ----------
    Fine:
      type: object
      properties:
        id:
          type: integer
          format: int64
        loanId:
          type: integer
          format: int64
        userName:
          type: string
          example: Bruno Bibliotecario
        value:
          type: number
          format: double
          example: 8.00
        daysLate:
          type: integer
          format: int32
          example: 4
        paymentStatus:
          $ref: '#/components/schemas/FinePaymentStatus'
        paymentDate:
          type: string
          format: date
          nullable: true
    FinePage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Fine'
        meta:
          $ref: '#/components/schemas/PageMeta'

    # ---------- Erro ----------
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: O campo 'email' e obrigatorio
        path:
          type: string
          example: /api/v1/users